app/requirements.txt
PyJWT
passlib[bcrypt]
pydantic[email]
